RESTO_CANTEEN_MIS.cpp
PROJECT DESCRIPTION:
This project includes some facilities of customers and products search, display, modification, delete etc. This software searches the client data which is store in the record and then gives permission to place the order by the customer. The software used for small scale canteens which have limited customers and products and wants to maintain the records.It generates invoice that is inclusive of all taxes (Set as per GST, India). It also provides user authentication option.

ADVANTAGES:

 1.Customer does not have to wait in long queues.
 2.It is easy to handle customers records and products records for future use.
 3.This software saves the time.
 4.Provides username authentication for individual users.
 5.Generates proper invoice.


DISADVANTAGES:

1.This system suitable for only small canteens.
2.Online facility is not available.Only CASH Payment feature available.
3.Product menu set as per Food Canteen.